--- Content.Server/Disease/Components/DiseaseVaccineCreatorComponent.cs
@@ -1,10 +1,0 @@
-namespace Content.Server.Disease.Components
-{
-    /// <summary>
-    /// Controls disease machine behavior specific to the
-    /// vaccine creating machine
-    /// </summary>
-    [RegisterComponent]
-    public sealed class DiseaseVaccineCreatorComponent : Component
-    {}
-}
