--- Content.Shared/Disease/SharedVaccineMachineComponent.cs
@@ -1,14 +1,0 @@
-using Robust.Shared.Serialization;
-
-namespace Content.Shared.Disease.Components
-{
-    [NetSerializable, Serializable]
-    public enum VaccineMachineUiKey : byte
-    {
-        Key,
-    }
-
-    [Serializable, NetSerializable]
-    public sealed class VaccineMachineBoundInterfaceState : BoundUserInterfaceState
-    {}
-}
